# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NCKH SaaS Multi-tenancy API
  version: 0.1.0
  description: >-
    Hợp đồng REST cho control plane và application plane. Các API nghiệp vụ lấy
    tenant từ host và JWT; không nhận tenant_id từ payload để phân quyền.
servers:
  - url: http://accounts.localhost:8080/api/v1
    description: Control host local
  - url: http://{tenant}.localhost:8080/api/v1
    description: Tenant host local
    variables:
      tenant:
        default: pool-demo
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Global session and available tenants
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LoginResponse'}
        default: {$ref: '#/components/responses/Error'}
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '200':
          description: New global account session with no tenant membership yet
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LoginResponse'}
        default: {$ref: '#/components/responses/Error'}
  /auth/tenant-transfer:
    post:
      tags: [Auth]
      operationId: createTenantTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TenantTransferRequest'}
      responses:
        '200':
          description: One-time transfer code
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TenantTransferResponse'}
        default: {$ref: '#/components/responses/Error'}
  /auth/exchange:
    post:
      tags: [Auth]
      operationId: exchangeTenantTransfer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExchangeRequest'}
      responses:
        '200':
          description: Tenant-bound access token; refresh and CSRF cookies are set
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TenantSessionResponse'}
        default: {$ref: '#/components/responses/Error'}
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshTenantSession
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: Rotated tenant session
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TenantSessionResponse'}
        default: {$ref: '#/components/responses/Error'}
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      security: []
      parameters:
        - $ref: '#/components/parameters/OptionalCsrfHeader'
      responses:
        '200': {description: Refresh session revoked}
        default: {$ref: '#/components/responses/Error'}
  /auth/me:
    get:
      tags: [Auth]
      operationId: currentUser
      responses:
        '200':
          description: Current user and optional tenant context
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MeResponse'}
        default: {$ref: '#/components/responses/Error'}
...
  /tenants/{tenantId}/onboarding:
    get:
      tags: [Tenants, Payments]
      operationId: getTenantOnboarding
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Tenant, payment and provisioning state visible to its Owner or Admin
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OnboardingView'}
        default: {$ref: '#/components/responses/Error'}
...
  /notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      responses:
        '200':
          description: Current user's notifications
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/NotificationView'}
        default: {$ref: '#/components/responses/Error'}
  /notifications/{notificationId}/read:
    patch:
      tags: [Notifications]
      operationId: markNotificationRead
      parameters:
        - name: notificationId
          in: path
          required: true
          schema: {$ref: '#/components/schemas/UUID'}
      responses:
        '200':
          description: Updated notification
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NotificationView'}
        default: {$ref: '#/components/responses/Error'}
  /notifications/read-all:
    post:
      tags: [Notifications]
      operationId: markAllNotificationsRead
      responses:
        '200': {description: All current user's notifications marked read}
        default: {$ref: '#/components/responses/Error'}
  /notifications/preferences:
    get:
      tags: [Notifications]
      operationId: getNotificationPreferences
      responses:
        '200':
          description: Notification preferences
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NotificationPreferences'}
        default: {$ref: '#/components/responses/Error'}
    put:
      tags: [Notifications]
      operationId: updateNotificationPreferences
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NotificationPreferences'}
      responses:
        '200':
          description: Updated notification preferences
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NotificationPreferences'}
        default: {$ref: '#/components/responses/Error'}
  /notifications/push-subscriptions:
    get:
      tags: [Notifications]
      operationId: listPushSubscriptions
      responses:
        '200':
          description: Current user's push subscriptions in this tenant
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PushSubscriptionView'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Notifications]
      operationId: addPushSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PushSubscriptionRequest'}
      responses:
        '200':
          description: Push subscription stored
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PushSubscriptionView'}
        default: {$ref: '#/components/responses/Error'}
  /notifications/push-subscriptions/{subscriptionId}:
    delete:
      tags: [Notifications]
      operationId: removePushSubscription
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema: {$ref: '#/components/schemas/UUID'}
      responses:
        '200': {description: Push subscription removed}
        default: {$ref: '#/components/responses/Error'}
...
  /tenants/{tenantId}/payment-session:
    post:
      tags: [Payments]
      operationId: createPaymentSession
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 8, maxLength: 120}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreatePaymentRequest'}
      responses:
        '200':
          description: Idempotent provider checkout session
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PaymentSessionView'}
        default: {$ref: '#/components/responses/Error'}
  /payments/webhooks/fake:
    post:
      tags: [Payments]
      operationId: handleFakePaymentWebhook
      security: []
      parameters:
        - name: X-Payment-Signature
          in: header
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FakePaymentWebhook'}
      responses:
        '200':
          description: Verified and idempotently processed callback
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PaymentResultView'}
        default: {$ref: '#/components/responses/Error'}
  /tenants/{tenantId}/payments/{paymentId}/fake-complete:
    post:
      tags: [Payments]
      operationId: completeFakePayment
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Local fake provider emits a signed, idempotent successful callback
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PaymentResultView'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string}
    OptionalCsrfHeader:
      name: X-CSRF-Token
      in: header
      required: false
      schema: {type: string}
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
  responses:
    Error:
      description: Structured API error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
  schemas:
    UUID: {type: string, format: uuid}
    Instant: {type: string, format: date-time}
    TenantRole: {type: string, enum: [OWNER, ADMIN, MEMBER]}
...
    TenantTier: {type: string, enum: [STARTER, PROFESSIONAL, ENTERPRISE]}
    TenantPlacement: {type: string, enum: [POOL, SILO_DATABASE]}
    TenantStatus: {type: string, enum: [PENDING_PAYMENT, PROVISIONING, ACTIVE, FAILED, SUSPENDED]}
    PaymentStatus: {type: string, enum: [CREATED, PENDING, SUCCEEDED, FAILED, EXPIRED]}
    ProvisioningStatus: {type: string, enum: [QUEUED, RUNNING, SUCCEEDED, RETRYABLE_FAILED, FAILED_ROLLED_BACK, ROLLBACK_FAILED]}
    PlatformRole: {type: string, enum: [SYSTEM_ADMIN]}
    ApiError:
      type: object
      required: [timestamp, status, code, message, path, fieldErrors]
      properties:
        timestamp: {$ref: '#/components/schemas/Instant'}
        status: {type: integer}
        code: {type: string}
        message: {type: string}
        path: {type: string}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        requestId: {type: [string, 'null']}
    UserView:
      type: object
      required: [id, email, displayName, platformRoles]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        email: {type: string, format: email}
        displayName: {type: string}
        platformRoles:
          type: array
          items: {$ref: '#/components/schemas/PlatformRole'}
    TenantView:
      type: object
      required: [id, slug, name, tier, status, placement, role]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        slug: {type: string}
        name: {type: string}
        tier: {$ref: '#/components/schemas/TenantTier'}
        status: {$ref: '#/components/schemas/TenantStatus'}
        placement: {$ref: '#/components/schemas/TenantPlacement'}
        role: {$ref: '#/components/schemas/TenantRole'}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
    RegisterRequest:
      type: object
      required: [email, displayName, password]
      properties:
        email: {type: string, format: email, maxLength: 320}
        displayName: {type: string, minLength: 2, maxLength: 160}
        password: {type: string, format: password, minLength: 10, maxLength: 128}
    LoginResponse:
      type: object
      required: [accessToken, expiresInSeconds, user, tenants]
      properties:
        accessToken: {type: string}
        expiresInSeconds: {type: integer, format: int64}
        user: {$ref: '#/components/schemas/UserView'}
        tenants:
          type: array
          items: {$ref: '#/components/schemas/TenantView'}
    TenantTransferRequest:
      type: object
      required: [tenantSlug]
      properties: {tenantSlug: {type: string}}
    TenantTransferResponse:
      type: object
      required: [code, redirectUrl, expiresInSeconds]
      properties:
        code: {type: string}
        redirectUrl: {type: string, format: uri}
        expiresInSeconds: {type: integer, format: int64}
    ExchangeRequest:
      type: object
      required: [code]
      properties: {code: {type: string}}
    TenantSessionResponse:
      type: object
      required: [accessToken, expiresInSeconds, user, activeTenant]
      properties:
        accessToken: {type: string}
        expiresInSeconds: {type: integer, format: int64}
        user: {$ref: '#/components/schemas/UserView'}
        activeTenant: {$ref: '#/components/schemas/TenantView'}
    MeResponse:
      type: object
      required: [user, tenantScoped]
      properties:
        user: {$ref: '#/components/schemas/UserView'}
        tenant: {anyOf: [{$ref: '#/components/schemas/TenantView'}, {type: 'null'}]}
        tenantScoped: {type: boolean}
...
    NotificationView:
      type: object
      required: [id, eventType, title, body, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        eventType: {type: string}
        title: {type: string}
        body: {type: string}
        readAt: {anyOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
        createdAt: {$ref: '#/components/schemas/Instant'}
    NotificationPreferences:
      type: object
      required: [inAppEnabled, emailEnabled, webPushEnabled]
      properties:
        inAppEnabled: {type: boolean}
        emailEnabled: {type: boolean}
        webPushEnabled: {type: boolean}
    PushSubscriptionRequest:
      type: object
      required: [endpoint, p256dh, auth]
      properties:
        endpoint: {type: string, format: uri}
        p256dh: {type: string}
        auth: {type: string}
    PushSubscriptionView:
      type: object
      required: [id, endpoint, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        endpoint: {type: string, format: uri}
        createdAt: {$ref: '#/components/schemas/Instant'}
...
    CreatePaymentRequest:
      type: object
      required: [amountMinor, currency, returnUrl]
      properties:
        amountMinor: {type: integer, format: int64, minimum: 1}
        currency: {type: string, pattern: '^[A-Za-z]{3}$'}
        returnUrl: {type: string, format: uri}
    FakePaymentWebhook:
      type: object
      required: [reference, eventId, status, amountMinor, currency]
      properties:
        reference: {type: string}
        eventId: {type: string}
        status: {type: string, enum: [SUCCEEDED, FAILED]}
        amountMinor: {type: integer, format: int64, minimum: 1}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
    PaymentSessionView:
      type: object
      required: [paymentId, provider, reference, checkoutUrl, status]
      properties:
        paymentId: {$ref: '#/components/schemas/UUID'}
        provider: {type: string}
        reference: {type: string}
        checkoutUrl: {type: string, format: uri}
        status: {$ref: '#/components/schemas/PaymentStatus'}
    PaymentResultView:
      type: object
      required: [paymentId, tenantId, status, duplicate]
      properties:
        paymentId: {$ref: '#/components/schemas/UUID'}
        tenantId: {$ref: '#/components/schemas/UUID'}
        status: {$ref: '#/components/schemas/PaymentStatus'}
        duplicate: {type: boolean}
    OnboardingPaymentView:
      type: object
      required: [id, provider, status, amountMinor, currency]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        provider: {type: string}
        status: {$ref: '#/components/schemas/PaymentStatus'}
        amountMinor: {type: integer, format: int64, minimum: 0}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
    OnboardingProvisioningView:
      type: object
      required: [id, status, attempts]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        status: {$ref: '#/components/schemas/ProvisioningStatus'}
        attempts: {type: integer, minimum: 0}
        lastErrorCode: {type: [string, 'null']}
        lastErrorMessage: {type: [string, 'null']}
    OnboardingView:
      type: object
      required: [tenant]
      properties:
        tenant: {$ref: '#/components/schemas/TenantView'}
        payment:
          anyOf: [{$ref: '#/components/schemas/OnboardingPaymentView'}, {type: 'null'}]
        provisioning:
          anyOf: [{$ref: '#/components/schemas/OnboardingProvisioningView'}, {type: 'null'}]
...